파일 snake type으로 rename
