+++ src/user/entity/user.entity.ts
@@ -7,14 +7,14 @@ import {
   OneToOne,
   PrimaryGeneratedColumn,
 } from "typeorm";
-import { PlateSetting } from "./platesetting.entity";
-import { UserTitle } from "./usertitle.entity";
-import { PatternProgress } from "./patternprogress.entity";
+import { PlateSetting } from "./plate-setting.entity";
+import { UserTitle } from "./user-title.entity";
+import { PatternProgress } from "./pattern-progress.entity";
 import { Post } from "src/post/entity/post.entity";
 import { Comment } from "src/comment/entity/comment.entity";
-import { PostReport } from "src/report/entity/postreport.entity";
-import { CommentReport } from "src/report/entity/commentreport.entity";
-import { UserReport } from "src/report/entity/userreport.entity";
+import { PostReport } from "src/report/entity/post-report.entity";
+import { CommentReport } from "src/report/entity/comment-report.entity";
+import { UserReport } from "src/report/entity/user-report.entity";
 
 @Entity()
 export class User {
